+++ R/utils_model_family.R
@@ -6,7 +6,7 @@ get_glm_family <- function(fit) {
 
   # do we have glm? if so, get link family. make exceptions
   # for specific models that don't have family function
-  if (any(mc %in% c("lme", "plm", "gls"))) {
+  if (any(mc %in% c("lme", "plm", "gls", "truncreg"))) {
     fitfam <- "gaussian"
     logit_link <- FALSE
     link.fun <- "identity"
@@ -15,10 +15,22 @@ get_glm_family <- function(fit) {
     fitfam <- faminfo@vfamily
     logit_link <- sjmisc::str_contains(faminfo@blurb, "logit")
     link.fun <- faminfo@blurb[3]
+  } else if (any(mc %in% c("zeroinfl", "hurdle"))) {
+    fitfam <- "negative binomial"
+    logit_link <- FALSE
+    link.fun <- NULL
+  } else if (any(mc %in% c("betareg"))) {
+    fitfam <- "beta"
+    logit_link <- fit$link$mean$name == "logit"
+    link.fun <- fit$link$mean$linkfun
+  } else if (any(mc %in% c("coxph"))) {
+    fitfam <- "survival"
+    logit_link <- TRUE
+    link.fun <- NULL
   } else {
     # "lrm"-object from pkg "rms" have no family method
     # so we construct a logistic-regression-family-object
-    if (mc == "lrm")
+    if (mc %in% c("lrm", "polr"))
       faminfo <- stats::binomial(link = "logit")
     else
       # get family info